+++ Example/iOS Example/AppDelegate.m
@@ -22,23 +22,16 @@
     application.appChannel = @"dev";
     application.appStoreID = @"12345678";
 
-    [application registerForRemoteNotifications];
+    [application registerForRemoteNotificationsWithCompletion:^(NSData * _Nullable deviceToken, NSError * _Nullable error) {
+        if (deviceToken) {
+            NSLog(@"%@", UIDevice.currentDevice.deviceTokenString);
+        }
+    }];
 
     NSLog(@"%@", application.analyticsInfo);
 
     [self.window makeKeyAndVisible];
     return YES;
-}
-
-- (void)application:(UIApplication *)application didRegisterForRemoteNotificationsWithDeviceToken:(NSData *)deviceToken
-{
-    UIDevice.currentDevice.deviceToken = deviceToken;
-    NSLog(@"device token: %@", deviceToken);
-}
-
-- (void)application:(UIApplication *)application didFailToRegisterForRemoteNotificationsWithError:(NSError *)error
-{
-    NSLog(@"%@", error);
 }
 
 @end